:recycle: api: moved sessionValidator to src/certification/session/domain/validators/
