+++ api/lib/application/sessions/session-controller.js
@@ -2,7 +2,7 @@ import { SessionPublicationBatchError } from '../http-errors.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { tokenService } from '../../domain/services/token-service.js';
 import * as sessionResultsLinkService from '../../domain/services/session-results-link-service.js';
-import * as sessionValidator from '../../domain/validators/session-validator.js';
+import * as sessionValidator from '../../../src/certification/session/domain/validators/session-validator.js';
 import * as events from '../../domain/events/index.js';
 import * as sessionSerializer from '../../infrastructure/serializers/jsonapi/session-serializer.js';
 import * as jurySessionSerializer from '../../infrastructure/serializers/jsonapi/jury-session-serializer.js';
