+++ api/src/certification/session/domain/validators/session-validator.js
@@ -1,12 +1,12 @@
 import BaseJoi from 'joi';
 import JoiDate from '@joi/date';
 const Joi = BaseJoi.extend(JoiDate);
-import { statuses } from '../models/Session.js';
-import { types } from '../models/CertificationCenter.js';
-import { CERTIFICATION_SESSIONS_ERRORS } from '../constants/sessions-errors.js';
-import { EntityValidationError } from '../errors.js';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
-import { CertificationVersion } from '../models/CertificationVersion.js';
+import { statuses } from '../../../../../lib/domain/models/Session.js';
+import { types } from '../../../../../lib/domain/models/CertificationCenter.js';
+import { CERTIFICATION_SESSIONS_ERRORS } from '../../../../../lib/domain/constants/sessions-errors.js';
+import { EntityValidationError } from '../../../../../lib/domain/errors.js';
+import { identifiersType } from '../../../../../lib/domain/types/identifiers-type.js';
+import { CertificationVersion } from '../../../../../lib/domain/models/CertificationVersion.js';
 
 const validationConfiguration = { abortEarly: false, allowUnknown: true };
