+++ api/src/certification/shared/domain/usecases/index.js
@@ -2,7 +2,7 @@ import { fileURLToPath } from 'node:url';
 import { dirname, join } from 'node:path';
 
 import * as sessionCodeService from '../../../../../lib/domain/services/session-code-service.js';
-import * as sessionValidator from '../../../../../lib/domain/validators/session-validator.js';
+import * as sessionValidator from '../../../session/domain/validators/session-validator.js';
 import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
 import * as sessionRepository from '../../../session/infrastructure/repositories/session-repository.js';
 import * as certificationCenterRepository from '../../../shared/infrastructure/repositories/certification-center-repository.js';
