+++ api/tests/certification/session/unit/domain/validators/session-validator_test.js
@@ -1,7 +1,7 @@
-import { expect, domainBuilder, catchErr } from '../../../test-helper.js';
-import { statuses } from '../../../../lib/domain/models/Session.js';
-import { EntityValidationError } from '../../../../lib/domain/errors.js';
-import * as sessionValidator from '../../../../lib/domain/validators/session-validator.js';
+import { expect, domainBuilder, catchErr } from '../../../../../test-helper.js';
+import { statuses } from '../../../../../../lib/domain/models/Session.js';
+import { EntityValidationError } from '../../../../../../lib/domain/errors.js';
+import * as sessionValidator from '../../../../../../src/certification/session/domain/validators/session-validator.js';
 
 const MISSING_VALUE = '';
